requirements.txt
Django==3.2.16
Pillow==9.3.0
drf-spectacular==0.24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2